wipdump

+++ game/system/newpad.h
@@ -91,6 +91,6 @@ void input_mode_pad_set(s64);
 
 void initialize();
 void update_gamepads();
-int rumble(int pad, float slow_motor, float fast_motor);
+int rumble(int pad, u16 slow_motor, u16 fast_motor);
 
 }  // namespace Pad
